purchase: read Tally Quantity as a decimal measure, add po_header.supplier_id
Tally's "Quantity" column is a decimal measure in kg, not a count of packs.
po_line.pack_count was INT and parser.py truncated it with int(), so every
fractional quantity lost its decimal and anything below 1 collapsed to 0 --
Rose Petals 20.8 -> 20, Labour Charges 0.96 -> 0. Roughly 12% of lines in
the Jan-Mar workbook carry a fractional quantity. The API schemas rejected
the decimal outright (int_from_float) and 500'd the whole preview.

- widen po_line.pack_count INT -> NUMERIC(15,3) (096; guarded on current
  type, so re-running is a no-op)
- PreviewLine.pack_count and PoLineOut.pack_count int -> float
- add po_header.supplier_id (097). The purchase code has always assumed it
  exists -- po_commit._HEADER_COLUMNS, po_query._EQUALITY_FIELDS,
  po_query._SORTABLE_COLUMNS and po_diff._HEADER_FIELDS all reference it --
  so every commit in a batch aborted with
  'column "supplier_id" of relation "po_header" does not exist'
- stop feeding uom from Tally's Col K. uom is the per-unit weight and comes
  from all_sku during enrichment; Col K is Tally's secondary quantity and is
  now surfaced separately as alt_units
- resolve parser columns through a header-name map instead of fixed
  positions, so a shifted column no longer reads SGST as total_amount

test_po_header_schema_drift pins the code's column lists against app/db/*.sql
so the schema cannot drift again. 107 tests pass.

+++ app/db/po_schema.sql
@@ -53,7 +53,9 @@ CREATE TABLE IF NOT EXISTS po_line (
     line_number             INT NOT NULL,
     sku_name                TEXT,
     uom                     TEXT,
-    pack_count              INT,
+    -- Tally's Quantity is a decimal measure (kg), not a count of packs.
+    -- See 097_po_line_pack_count_numeric.sql.
+    pack_count              NUMERIC(15,3),
     po_weight               NUMERIC(15,3),
     rate                    NUMERIC(15,3),
     amount                  NUMERIC(15,3),
